Apply gyro filter checks to accelerometer
In 'AP_InertialSensor_Backend::_notify_new_gyro_raw_sample(...)'
the result of the filters is checked before the value is used.
This is not the case in
'AP_InertialSensor_Backend::_notify_new_accel_raw_sample(...)'.
Here, the value is used in any case and only the filter is reset.
Therefore I applied the same check as for the gyros, meaning that
the filtered accel value is only used if the result is not
nan or inf.

DANGER:
- THIS HAS NOT BEEN SIMULATION/FLIGHT TESTED YET!
- I can not estimate if this change will bring problems!