Add reading function for last_raw_gyro
Added a reading function to the sensor frontend in
ArduPilot to read the value of _last_raw_gyro in
'AP_InertialSensor'. In this way the reading is the same
as for the raw accelerometer and it is no longer necessary
to apply the scaling factor to the gyro value.

Only downside could be that this value is not thread-safe,
because it is not updated by
'AP_InertialSensor_Backend::update_gyro/accel' but
'AP_InertialSensor_Backend::_notify_new_gyro_raw_sample'.
This would also apply to last_raw_accel!

A possible solution could be to add additional variables to
'AP_InertialSensor' and update these variables with
'AP_InertialSensor_Backend::update_gyro/accel'...