Add derivative of filtered gyro
Added a simple discrete derivative calculation to the sensor backend in
ArduPilot. It calculates the two-point backward finite difference using
the filtered gyro signal and writes it to the sensor frontend. The
derivative of the filtered gyro signal can be used in Simulink via
'omega_Kb_f_dt'.

Also added the filtered gyro signal to the Simulink interface, which
can be used via 'omega_Kb_f'.

The name of the raw gyro signal in Simulink was changed from 'omega_Kb'
to 'omega_Kb_raw'.